backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
PyPDF2==3.0.1
python-docx==1.1.2
pypdf==5.1.0

